add goodreads into theme if plugin exists
as a mock-sticky post for Status cpt. done as a template part for
new component like design model.

stickies are not that great for CPTs, this method was quicker and can be
eaisly moved/toggled elsewhere in future.

it should be concidered to place this in the plugin-dsca as this will be used
in the future